Restore default dtype to float32 (mxnet nd standard) and fix dtype and ctx bugs in model_io

+++ python/mxmodules/convolution.py
@@ -20,7 +20,7 @@ from module import Module
 
 class Convolution(Module):
 
-    def __init__(self, filtersize=(5,5,3,32), stride = (2,2), ctx=mx.cpu(), dtype='float64'):
+    def __init__(self, filtersize=(5,5,3,32), stride = (2,2), ctx=mx.cpu(), dtype='float32'):
         '''
         Constructor for a Convolution layer.
